move test to test_custom_primitives.py
